+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Adding `xan sort --columns`.
 * Adding `xan view -T/--tee`.
 * Adding `now`, `fractional_days`, `to_timezone`, `to_local_timezone`, `with_timezone`, `with_local_timezone`, `without_timezone`, `to_timestamp`, `to_timestamp_ms`, `from_timestamp`, `from_timestamp_ms`, `span`, `date` & `time` moonblade functions.
+* Better type inference with `xan stats`, and the `type` & `types` aggregation functions, now including more types for temporal values (`zoned_datetime`, `datetime`, `date` & `time`).
 
 *Fixes*
 
@@ -39,7 +40,6 @@
 *Quality of Life*
 
 * Adding proper help to `xan heatmap`.
-* Better type inference with `xan stats`, and the `type` & `types` aggregation functions.
 
 ## 0.56.0